+++ src/exceptions.mjs
@@ -1,2 +1,3 @@
 export { default as PHPJSError } from './exceptions/PHPJSError.mjs';
 export { default as TypePHPJSError } from './exceptions/TypePHPJSError.mjs';
+export { default as ValuePHPJSError } from './exceptions/ValuePHPJSError.mjs';