+++ tests/exceptions.global.test.mjs
@@ -4,4 +4,5 @@ import '../src/exceptions.global.js';
 test('exceptions.global', () => {
     expect(PHPJSError).not.toBeUndefined();
     expect(TypePHPJSError).not.toBeUndefined();
+    expect(ValuePHPJSError).not.toBeUndefined();
 });