+++ tests/exceptions.test.mjs
@@ -1,7 +1,8 @@
 import { expect, test } from '@jest/globals';
-import { PHPJSError, TypePHPJSError } from '../src/exceptions.mjs';
+import { PHPJSError, TypePHPJSError, ValuePHPJSError } from '../src/exceptions.mjs';
 
 test('exceptions', () => {
     expect(PHPJSError).not.toBeUndefined();
     expect(TypePHPJSError).not.toBeUndefined();
+    expect(ValuePHPJSError).not.toBeUndefined();
 });